+++ src/language/lexer/lexer.c
@@ -1069,9 +1069,9 @@ convert_numeric_string_to_char_string (struct lexer *lexer,
 
   byte_cnt = ds_length (&lexer->tokstr) / chars_per_byte;
   if (ds_length (&lexer->tokstr) % chars_per_byte)
-    msg (SE, _("String of %s digits has %d characters, which is not a "
+    msg (SE, _("String of %s digits has %zu characters, which is not a "
 	       "multiple of %d."),
-	 base_name, (int) ds_length (&lexer->tokstr), chars_per_byte);
+	 base_name, ds_length (&lexer->tokstr), chars_per_byte);
 
   p = ds_cstr (&lexer->tokstr);
   for (i = 0; i < byte_cnt; i++)
@@ -1207,8 +1207,8 @@ finish:
 
   if (ds_length (&lexer->tokstr) > 255)
     {
-      msg (SE, _("String exceeds 255 characters in length (%d characters)."),
-	   (int) ds_length (&lexer->tokstr));
+      msg (SE, _("String exceeds 255 characters in length (%zu characters)."),
+	   ds_length (&lexer->tokstr));
       ds_truncate (&lexer->tokstr, 255);
     }
